requirements.txt
Django==5.0.1
requests==2.31.0
python-dotenv==1.0.0
django-crispy-forms==2.1
crispy-bootstrap4==2023.1